requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
Django==3.2.6
django-debug-toolbar==3.2.2
flake8==3.9.2
isort==5.9.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
regex==2021.8.3
sentry-sdk==1.3.1
sqlparse==0.4.1
tomli==1.2.0
urllib3==1.26.6
